+++ pybamm/models/submodels/porosity/base_porosity.py
@@ -23,6 +23,9 @@ class BaseModel(pybamm.BaseSubModel):
         self, eps_n, eps_s, eps_p, set_leading_order=False
     ):
 
+        eps_n_av = pybamm.x_average(eps_n)
+        eps_s_av = pybamm.x_average(eps_s)
+        eps_p_av = pybamm.x_average(eps_p)
         eps = pybamm.Concatenation(eps_n, eps_s, eps_p)
 
         variables = {
@@ -30,9 +33,9 @@ class BaseModel(pybamm.BaseSubModel):
             "Negative electrode porosity": eps_n,
             "Separator porosity": eps_s,
             "Positive electrode porosity": eps_p,
-            "X-averaged negative electrode porosity": pybamm.x_average(eps_n),
-            "X-averaged separator porosity": pybamm.x_average(eps_s),
-            "X-averaged positive electrode porosity": pybamm.x_average(eps_p),
+            "X-averaged negative electrode porosity": eps_n_av,
+            "X-averaged separator porosity": eps_s_av,
+            "X-averaged positive electrode porosity": eps_p_av,
         }
 
         # activate material volume fractions
@@ -66,6 +69,9 @@ class BaseModel(pybamm.BaseSubModel):
         self, deps_n_dt, deps_s_dt, deps_p_dt, set_leading_order=False
     ):
 
+        deps_n_dt_av = pybamm.x_average(deps_n_dt)
+        deps_s_dt_av = pybamm.x_average(deps_s_dt)
+        deps_p_dt_av = pybamm.x_average(deps_p_dt)
         deps_dt = pybamm.Concatenation(deps_n_dt, deps_s_dt, deps_p_dt)
 
         variables = {
@@ -73,25 +79,19 @@ class BaseModel(pybamm.BaseSubModel):
             "Negative electrode porosity change": deps_n_dt,
             "Separator porosity change": deps_s_dt,
             "Positive electrode porosity change": deps_p_dt,
-            "X-averaged porosity change": pybamm.x_average(deps_dt),
-            "X-averaged negative electrode porosity change": pybamm.x_average(
-                deps_n_dt
-            ),
-            "X-averaged separator porosity change": pybamm.x_average(deps_s_dt),
-            "X-averaged positive electrode porosity change": pybamm.x_average(
-                deps_p_dt
-            ),
+            "X-averaged negative electrode porosity change": deps_n_dt_av,
+            "X-averaged separator porosity change": deps_s_dt_av,
+            "X-averaged positive electrode porosity change": deps_p_dt_av,
         }
 
         if set_leading_order is True:
             variables.update(
                 {
                     "Leading-order x-averaged "
-                    + "negative electrode porosity change": pybamm.x_average(deps_n_dt),
+                    + "negative electrode porosity change": deps_n_dt_av,
+                    "Leading-order x-averaged separator porosity change": deps_s_dt_av,
                     "Leading-order x-averaged "
-                    + "separator porosity change": pybamm.x_average(deps_s_dt),
-                    "Leading-order x-averaged "
-                    + "positive electrode porosity change": pybamm.x_average(deps_p_dt),
+                    + "positive electrode porosity change": deps_p_dt_av,
                 }
             )